+++ irc3/plugins/core.py
@@ -30,10 +30,24 @@ class Core(object):
         self.max_lag = int(self.bot.config.get('max_lag'))
         self.reconn_handle = None
         self.ping_handle = None
+        self.events = (
+            event(r"^:(?P<srv>\S+) 005 (?P<me>\S+) (?P<data>.+) :\S+",
+                  self.set_config),
+            event(rfc.CONNECTED, self.detach_events),
+        )
 
     def connection_made(self):
+        # handle server config
+        self.bot.config['server_config'] = {}
+        self.detach_events()
+        self.bot.attach_events(insert=True, *self.events)
+
+        # ping/ping
         self.connection_made_at = self.bot.loop.time()
         self.pong(event='CONNECT', data='')
+
+    def detach_events(self, **kwargs):
+        self.bot.detach_events(*self.events)
 
     def reconnect(self):  # pragma: no cover
         self.bot.log.info(
@@ -75,3 +89,14 @@ class Core(object):
             self.bot.set_nick(self.bot.nick + '_')
         self.bot.log.debug('Trying to regain nickname in 30s...')
         self.bot.loop.call_later(30, self.bot.set_nick, self.bot.original_nick)
+
+    def set_config(self, data=None, **kwargs):
+        """Store server config"""
+        config = self.bot.config['server_config']
+        for opt in data.split(' '):
+            if '=' in opt:
+                opt, value = opt.split('=', 1)
+            else:
+                value = True
+            if opt.isupper():
+                config[opt] = value