+++ irc3/rfc.py
@@ -14,11 +14,16 @@ class raw(str):
         r.re = regexp
         return r
 
+CONNECTED = raw.new('CONNECTED',
+                    r'^:(?P<srv>\S+) (376|422) (?P<me>\S+) :(?P<data>.*)')
+
 PING = raw.new('PING', r'PING :(?P<data>.*)')
 PONG = raw.new(
     'PONG',
     r':(?P<server>\S+) PONG (?P=server) :(?P<data>.*)')
+
 NEW_NICK = raw.new('NEW_NICK', r':(?P<nick>\S+) NICK (?P<new_nick>\S+)')
+
 JOIN = raw.new('JOIN', r':(?P<mask>\S+) JOIN (?P<channel>\S+)')
 PART = raw.new('PART',
                r':(?P<mask>\S+) PART (?P<channel>\S+)(\s+:(?P<data>.*)|$)')